wrote the amazon repository

+++ app/src/main/java/com/kev/windowshopper/model/Product.kt
@@ -11,5 +11,5 @@ data class Product(
     val productLink: String,
     val productPrice: String,
     val productRating: Float,
-    val totalReviews: Float
+    val totalReviews: Int
 )
